Explain installation instructions (#349)
* Stop recommending adoptopenjdk

It is not notarized but installed with quarantine flags. As a result,
it does not work out of the box. And because it is maintained by
Homebrew-Casks and we cannot agree with them how to package/distribute
it, we cannot easily fix it. Our tap does not have that problem.

* Explain installation instructions